+++ eras/alonzo/impl/test/Test/Cardano/Ledger/Alonzo/Imp/TxInfoSpec.hs
@@ -48,7 +48,7 @@ spec = withImpInit @(LedgerSpec AlonzoEra) $ describe "TxInfo" $ do
             , ltiSystemStart = systemStart
             , ltiUTxO = utxo
             , ltiTx = tx
-            , ltiMemoizedSubTransactions = mempty
+            , ltiLevelInfo = ()
             }
       void $ expectRight $ unPlutusTxInfoResult $ toPlutusTxInfo SPlutusV1 lti
     it "toPlutusTxInfo does not fail when Byron scripts are present in TxIns" $ do
@@ -76,6 +76,6 @@ spec = withImpInit @(LedgerSpec AlonzoEra) $ describe "TxInfo" $ do
             , ltiSystemStart = systemStart
             , ltiUTxO = utxo
             , ltiTx = tx
-            , ltiMemoizedSubTransactions = mempty
+            , ltiLevelInfo = ()
             }
       void $ expectRight $ unPlutusTxInfoResult $ toPlutusTxInfo SPlutusV1 lti