+++ eras/babbage/impl/testlib/Test/Cardano/Ledger/Babbage/Imp.hs
@@ -13,10 +13,12 @@ module Test.Cardano.Ledger.Babbage.Imp (
   babbageOnlySpec,
 ) where
 
+import Cardano.Ledger.Alonzo.Plutus.Context (EraPlutusContext (..))
 import Cardano.Ledger.Babbage.Core
 import Cardano.Ledger.Babbage.State
 import qualified Cardano.Ledger.Shelley.Rules as Shelley
 import Control.State.Transition (Event)
+import Data.Default (Default)
 import qualified Test.Cardano.Ledger.Alonzo.Imp as Alonzo
 import Test.Cardano.Ledger.Alonzo.ImpTest
 import qualified Test.Cardano.Ledger.Babbage.Imp.PoolSpec as POOL
@@ -29,6 +31,7 @@ import Test.Cardano.Ledger.Imp.Common
 spec ::
   ( BabbageEraImp era
   , Event (EraRule "RUPD" era) ~ Shelley.RupdEvent
+  , Default (LevelTxInfo TopTx era)
   ) =>
   proxy era ->
   Spec