+++ eras/conway/impl/src/Cardano/Ledger/Conway/TxInfo.hs
@@ -155,6 +155,7 @@ import qualified PlutusLedgerApi.V3.MintValue as PV3
 
 instance EraPlutusContext ConwayEra where
   type ContextError ConwayEra = ConwayContextError ConwayEra
+  type LevelTxInfo level ConwayEra = ()
 
   data TxInfoResult ConwayEra
     = ConwayTxInfoResult -- Fields must be kept lazy